url_shortener/requirements.txt
Flask
Redis
fastwsgi
cassandra-driver
